+++ PluginFramework/src/main/java/ws/argo/plugin/transport/responder/TransportConfigException.java
@@ -1,15 +1,31 @@
+/*
+ * Copyright 2015 Jeff Simpson.
+ *
+ * Licensed under the MIT License, (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://opensource.org/licenses/MIT
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package ws.argo.plugin.transport.responder;
 
 /**
- * This exception gets thrown when some kind of error occured during the
- * configuration of a Transport.
+ * TransportConfigException is an exception that can occur if the Transport is
+ * provided bad configuration parameters.
  * 
  * @author jmsimpson
  *
  */
 public class TransportConfigException extends Exception {
 
-  private static final long serialVersionUID = 753487924802419379L;
+  private static final long serialVersionUID = 4913095738616220364L;
 
   public TransportConfigException() {
     super();
@@ -22,4 +38,5 @@ public class TransportConfigException extends Exception {
   public TransportConfigException(String message) {
     super(message);
   }
+
 }
