+++ Responder/ResponderDaemon/src/main/java/ws/argo/responder/Responder.java
@@ -48,9 +48,9 @@ import org.joda.time.Instant;
 
 import ws.argo.plugin.probehandler.ProbeHandlerConfigException;
 import ws.argo.plugin.probehandler.ProbeHandlerPluginIntf;
-import ws.argo.responder.transport.ProbeProcessor;
-import ws.argo.responder.transport.Transport;
-import ws.argo.responder.transport.TransportConfigException;
+import ws.argo.plugin.transport.responder.ProbeProcessor;
+import ws.argo.plugin.transport.responder.Transport;
+import ws.argo.plugin.transport.responder.TransportConfigException;
 import ws.argo.wireline.probe.ProbeWrapper;
 
 /**
