+++ Responder/ResponderDaemon/src/main/java/ws/argo/responder/transport/AmazonSNSTransport.java
@@ -22,6 +22,9 @@ import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.sns.AmazonSNSClient;
 import com.amazonaws.services.sns.model.SubscribeRequest;
 
+import ws.argo.plugin.transport.responder.ProbeProcessor;
+import ws.argo.plugin.transport.responder.Transport;
+import ws.argo.plugin.transport.responder.TransportConfigException;
 import ws.argo.responder.Responder;
 import ws.argo.responder.transport.sns.SNSListener;
