+++ Responder/ResponderDaemon/src/main/java/ws/argo/responder/transport/MulticastTransport.java
@@ -15,6 +15,9 @@ import java.util.Properties;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import ws.argo.plugin.transport.responder.ProbeProcessor;
+import ws.argo.plugin.transport.responder.Transport;
+import ws.argo.plugin.transport.responder.TransportConfigException;
 import ws.argo.responder.Responder;
 import ws.argo.wireline.probe.ProbeParseException;
 import ws.argo.wireline.probe.ProbeWrapper;
